IA-64 work. Redefine FRAME_SIZE to "(16*1024)" instead of "2*4096". Note that the latter definition was wrong because of the lack of "()". Take panic_printf(), cpu_halt() and memsetb() out of fake.s and implement them properly in asm.S. Move ski_init_console() from arch_post_mm_init() to arch_pre_mm_init(). In start.S, write r0 to ar.bspstore.
Janitorial fixes.
Add count_t type.
Replace several occurrences of int and __u32 by count_t or size_t.
Shorten several panic strings.

+++ include/mm/heap.h
@@ -40,7 +40,7 @@ struct chunk {
 	__u8 data[0];
 };
 
-extern void heap_init(__address heap, __u32 size);
+extern void heap_init(__address heap, size_t size);
 
 extern void *malloc(size_t size);
 extern void free(void *ptr);
